requirements.txt
Flask
requests
beautifulsoup4
nltk
spacy
pandas
PyPDF2
scikit-learn
python-docx